# BackEnd/docs/swagger.yaml
basePath: /
definitions:
  account.Account:
    properties:
      bank:
        example: Bank of America
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: acc_123456789
        type: string
      initial_balance:
        example: 1000.5
        type: number
      name:
        example: My Savings Account
        type: string
      user_id:
        example: user_987654321
        type: string
    type: object
...
  dto.AccountRequest:
    properties:
      bank:
        example: Bank of America
        maxLength: 100
        minLength: 2
        type: string
      initial_balance:
        example: 1000.5
        maximum: 9.9999999999e+08
        minimum: 0
        type: number
      name:
        example: My Savings Account
        maxLength: 100
        minLength: 2
        type: string
    required:
    - bank
    - initial_balance
    - name
    type: object
  dto.AccountResponse:
    properties:
      account_info:
        $ref: '#/definitions/account.Account'
      current_balance:
        example: 1250.75
        type: number
    type: object
  dto.AccountUpdateRequest:
    properties:
      bank:
        example: Updated Bank Name
        maxLength: 100
        minLength: 2
        type: string
      name:
        example: Updated Savings Account
        maxLength: 100
        minLength: 2
        type: string
    required:
    - bank
    - name
    type: object
  dto.BudgetRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
    type: object
  dto.BudgetResponse:
    properties:
      amount:
        type: number
      category_id:
        type: string
      created_at:
        type: string
      current_amount:
        type: number
      id:
        type: string
      user_id:
        type: string
    type: object
  dto.CategoryRequest:
    properties:
      color:
        example: '#FF6B6B'
        maxLength: 7
        minLength: 4
        type: string
      icon:
        example: "\U0001F354"
        maxLength: 50
        minLength: 1
        type: string
      name:
        example: Food & Dining
        maxLength: 50
        minLength: 2
        type: string
    required:
    - color
    - icon
    - name
    type: object
  dto.CategoryResponse:
    properties:
      color:
        example: '#FF6B6B'
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      icon:
        example: "\U0001F354"
        type: string
      id:
        example: cat_123456789
        type: string
      name:
        example: Food & Dining
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gestordePresupuesto.com
    name: API Support
    url: http://github.com/osmait/gestorDePresupuesto
  description: API for managing personal budget, transactions, accounts, and investments
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Budget Management System API
  version: "1.0"
paths:
  /account:
    get:
      consumes:
      - application/json
      description: Retrieve all bank accounts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of user accounts
          schema:
            items:
              $ref: '#/definitions/dto.AccountResponse'
            type: array
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Get all user accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Create a new bank account for the authenticated user
      parameters:
      - description: Account creation data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.AccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
        "400":
          description: Bad request - Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Create a new account
      tags:
      - Accounts
  /account/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific bank account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Delete an account
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      description: Update name and bank information for a specific account owned by
        the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.AccountUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found or not owned by user
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Update an existing account
      tags:
      - Accounts
  /budget:
    get:
      consumes:
      - application/json
      description: Retrieve all budget limits for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of user budgets
          schema:
            items:
              $ref: '#/definitions/dto.BudgetResponse'
            type: array
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Get all user budgets
      tags:
      - Budgets
    post:
      consumes:
      - application/json
      description: Create a new budget limit for a specific category for the authenticated
        user
      parameters:
      - description: Budget creation data
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/dto.BudgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Budget created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Create a new budget
      tags:
      - Budgets
  /budget/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific budget by ID
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Budget not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Delete a budget
      tags:
      - Budgets
  /category:
    get:
      consumes:
      - application/json
      description: Retrieve all categories for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of user categories
          schema:
            items:
              $ref: '#/definitions/dto.CategoryResponse'
            type: array
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Get all user categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new expense/income category for the authenticated user
      parameters:
      - description: Category creation data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Create a new category
      tags:
      - Categories
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Delete a category
      tags:
      - Categories
...
securityDefinitions:
  JWT:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"